refactor: extract persist restore helpers
